feat: 상품 관리 및 쿠폰 관리 컴포넌트 분리
