+++ src/basic/components/index.ts
@@ -4,3 +4,5 @@ export { ProductList } from './ProductList';
 export { Cart } from './Cart';
 export { NotificationList } from './NotificationList';
 export { AdminPanel } from './AdminPanel';
+export { ProductManagement } from './ProductManagement';
+export { CouponManagement } from './CouponManagement';
